+++ create-module.js
@@ -1,7 +1,7 @@
 const shell = require("shelljs");
 const fs = require("fs");
 const chalk = require("chalk");
-const { camelCase } = require("change-case");
+const { camelCase, paramCase } = require("change-case");
 shell.config.execPath = String(shell.which("node"));
 
 const {
@@ -12,7 +12,13 @@ const {
   createSaga,
   createConstants,
   createActions,
-  createReducer
+  createReducer,
+  createTestsActions,
+  createTestsReducer,
+  createTestsSaga,
+  createTestsIndex,
+  createTestsPage,
+  createTestsComponent,
 } = require("./templates");
 
 const findImportIndex = data =>
@@ -61,6 +67,20 @@ const createFunctionalComponent = (
   shell.echo(chalk.blue(`${name} component created.`));
 };
 
+const createComponentTest = (name, cwd = "", moduleName) => {
+  const fileName = `${name}.test.js`;
+  shell.exec(`touch ${fileName}`, { cwd });
+  const content = createTestsComponent(moduleName, name);
+  writeFile(cwd + fileName, content);
+}
+
+const createPageTest = (name, cwd = "", moduleName) => {
+  const fileName = `${name}.test.js`;
+  shell.exec(`touch ${fileName}`, { cwd });
+  const content = createTestsPage(moduleName, name);
+  writeFile(cwd + fileName, content);
+}
+
 const updateRootReducer = (config, name) => {
   insertDataInFile(`${config.workspace}${config.rootReducersPath}`, [
     {
@@ -92,7 +112,7 @@ const updateRoutes = (config, name, pages) => {
   importString = '',
   routeString = '';
   pages.forEach(page => {
-    pageString += `    ${page}Page: '/${name}/${page}/',\n`;
+    pageString += `    ${page}Page: '/${paramCase(name)}/${paramCase(page)}/',\n`;
     importString += `import ${page} from '${name}/Pages/${page}';\n`;
     routeString += `        <Route path={routesPath.${page}Page} component={${page}} />\n`;
   });
@@ -100,19 +120,20 @@ const updateRoutes = (config, name, pages) => {
     {
       indexCallBack: data => data.indexOf("\n", data.lastIndexOf(",")) + 1,
       str: `\n    /* ${name} Page */
-    ${name}Page: '/${name}/',\n${pageString}`
+    ${name}Page: '/${paramCase(name)}/',\n${pageString}`
     }
   ]);
   insertDataInFile(`${config.workspace}${config.routes}`, [
     {
       indexCallBack: findImportIndex,
-      str: `import ${name} from '${name}';
+      str: `\n /* ${name} Page */
+import ${name} from '${name}';
 ${importString}`
     },
     {
-      indexCallBack: data => data.indexOf("\n", data.lastIndexOf("/>")) + 1,
+      indexCallBack: data => data.indexOf("\n", data.lastIndexOf("/>", data.indexOf("Unused Currently"))) + 1,
       str: `\n        {/* ${name} */}
-        <Route path={routesPath.${name}Page} component={${name}} />\n${routeString}`
+        <Route exact path={routesPath.${name}Page} component={${name}} />\n${routeString}`
     }
   ]);
 }
@@ -120,6 +141,10 @@ ${importString}`
 const createModule = (config, name, components = [], pages = []) => {
   const modulePath = `${config.modulePath}${name}/`;
   const componentsPath = `${modulePath}Components/`;
+  const testsPath = `${modulePath}__tests__/`;
+  const testsMockPath = `${modulePath}__tests__/__mocks__/`;
+  const testsComponentsPath = `${modulePath}__tests__/__components__/`;
+  const testsPagesPath = `${modulePath}__tests__/__pages__/`;
   const pagesPath = `${modulePath}Pages/`;
   shell.exec(`rm -rf ${modulePath}`);
   shell.exec(`mkdir ${modulePath}`);
@@ -135,10 +160,20 @@ const createModule = (config, name, components = [], pages = []) => {
   createModuleActionsJs(`${modulePath}actions.js`);
   shell.exec(`touch ${modulePath}reducer.js`);
   createModuleReducerJs(`${modulePath}reducer.js`);
+  shell.exec(`rm -rf ${testsPath}`);
+  shell.exec(`mkdir ${testsPath}`);
+  shell.exec(`rm -rf ${testsMockPath}`);
+  shell.exec(`mkdir ${testsMockPath}`);
+  createModuleTestsActionsJs(`${testsPath}actions.test.js`, name);
+  createModuleTestsReducerJs(`${testsPath}reducer.test.js`, name);
+  createModuleTestsSagaJs(`${testsPath}saga.test.js`, name);
+  createModuleTestsIndexJs(`${testsPath}index.test.js`, name);
   if (components.length) {
     shell.exec(`rm -rf ${componentsPath}`);
     shell.exec(`mkdir ${componentsPath}`);
-    components.forEach(component =>
+    shell.exec(`rm -rf ${testsComponentsPath}`);
+    shell.exec(`mkdir ${testsComponentsPath}`);
+    components.forEach(component => {
       createFunctionalComponent(
         component,
         createComponentIndexJs,
@@ -146,12 +181,15 @@ const createModule = (config, name, components = [], pages = []) => {
         componentsPath,
         name
       )
-    );
+      createComponentTest(component, testsComponentsPath, name);
+      });
   }
   if (pages.length) {
     shell.exec(`rm -rf ${pagesPath}`);
     shell.exec(`mkdir ${pagesPath}`);
-    pages.forEach(page =>
+    shell.exec(`rm -rf ${testsPagesPath}`);
+    shell.exec(`mkdir ${testsPagesPath}`);
+    pages.forEach(page => {
       createFunctionalComponent(
         page,
         createPageIndexJs,
@@ -159,7 +197,8 @@ const createModule = (config, name, components = [], pages = []) => {
         pagesPath,
         name
       )
-    );
+      createPageTest(page, testsPagesPath, name);
+      });
   }
   updateRootReducer(config, name);
   shell.echo(chalk.green(`${name} Reducer added in rootReducer.`));
@@ -220,6 +259,26 @@ const createModuleReducerJs = path => {
   const content = createReducer();
   writeFile(path, content);
 };
+
+const createModuleTestsActionsJs = (path, name) => {
+  const content = createTestsActions(name);
+  writeFile(path, content);
+}
+
+const createModuleTestsReducerJs = (path, name) => {
+  const content = createTestsReducer(name);
+  writeFile(path, content);
+}
+
+const createModuleTestsSagaJs = (path, name) => {
+  const content = createTestsSaga(name);
+  writeFile(path, content);
+}
+
+const createModuleTestsIndexJs = (path, name) => {
+  const content = createTestsIndex(name);
+  writeFile(path, content);
+}
 
 module.exports = {
   createModule,