+++ templates/constants.js
@@ -1,11 +1,7 @@
 module.exports = {
-    createConstants: () => `const SHOW_BACK_BUTTON = 'SHOW_BACK_BUTTON';
-const CLEAR_STATE = 'CLEAR_STATE';
-const UPDATE_LAYOUT_PROPS = 'UPDATE_LAYOUT_PROPS';
+    createConstants: () => `const CLEAR_STATE = 'CLEAR_STATE';
 
 export {
-    SHOW_BACK_BUTTON,
     CLEAR_STATE,
-    UPDATE_LAYOUT_PROPS,
 };`
 }