+++ templates/page.js
@@ -1,14 +1,15 @@
+const { camelCase } = require('change-case');
 module.exports = {
-  createPage: (name, moduleName) => `import React, { useEffect } from 'react';
+  createPage: (name, moduleName) => `import React, { useEffect, useCallback } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import _ from 'lodash';
 import { buildUrl } from 'utils';
 import { routesPath } from 'Routes';
-import { updateLayoutProps, showBackButton, clearState } from '${moduleName}/actions';
+import { updateLayoutProps, showBackButton } from 'helpers/dispatchAction'
+import { clearState } from '${moduleName}/actions';
 import styles from './index.scss';
 
-const ${name} = ({ history, match, updateLayoutProps, showBackButton, clearState }) => {
+const ${name} = ({ history, match, clearState }) => {
     useEffect(() => {
         updateLayoutProps({
             whiteHeader: true,
@@ -22,7 +23,7 @@ const ${name} = ({ history, match, updateLayoutProps, showBackButton, clearState
         });
         return () => clearState();
     }, []);
-    const handleBackButtonClick = () => ({});
+    const handleBackButtonClick = useCallback(() => ({}), []);
     return (
         <div className={styles.container}>
             ${name} page.
@@ -33,17 +34,13 @@ const ${name} = ({ history, match, updateLayoutProps, showBackButton, clearState
 ${name}.propTypes = {
     history: PropTypes.object.isRequired,
     match: PropTypes.object.isRequired,
-    updateLayoutProps: PropTypes.func.isRequired,
-    showBackButton: PropTypes.func.isRequired,
     clearState: PropTypes.func.isRequired
 };
 
-const mapStateToProps = state => ({
+const mapStateToProps = ({ ${camelCase(name)} }) => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-    updateLayoutProps: payload => dispatch(updateLayoutProps(payload)),
-    showBackButton: payload => dispatch(showBackButton(payload)),
     clearState: () => dispatch(clearState())
 });
 
@@ -51,13 +48,5 @@ export default connect(mapStateToProps, mapDispatchToProps)(${name});`,
   createPageScss: () => `.rhsContent {
     max-width: 100% !important;
     padding: 20px 0px 40px 0px !important;
-}
-
-.container{
-
-}
-
-.loaderContainer{
-
 }`
 };