+++ templates/reducer.js
@@ -1,8 +1,5 @@
 module.exports = {
     createReducer: () => `import {
-    SOME_CONSTANT,
-    SOME_CONSTANT_SUCCESS,
-    SOME_CONSTANT_ERROR,
     CLEAR_STATE
 } from './constants';
 
@@ -14,12 +11,6 @@ const initState = {
 
 export default function(state = initState, { type, payload }) {
     switch(type){
-        case SOME_CONSTANT:
-            return { ...state, loading: true, error: {} };
-        case SOME_CONSTANT_SUCCESS:
-            return { ...state, loading: false, data: payload };
-        case SOME_CONSTANT_ERROR:
-            return { ...state, loading: false, error: payload };
         case CLEAR_STATE:
             return initState;
         default: